v2 of integration prototype. Fully functional
