+++ src/integration-test/groovy/org/pih/warehouse/component/api/product/CategoryApiWrapper.groovy
@@ -1,0 +1,34 @@
+package org.pih.warehouse.component.api.product
+
+import io.restassured.path.json.JsonPath
+import io.restassured.specification.RequestSpecification
+
+import org.pih.warehouse.component.api.base.ApiWrapper
+
+class CategoryApiWrapper extends ApiWrapper {
+
+    CategoryApi categoryApi
+
+    CategoryApiWrapper(RequestSpecification defaultRequestSpec) {
+        super(defaultRequestSpec)
+
+        categoryApi = new CategoryApi(defaultRequestSpec)
+    }
+
+    JsonPath listOK() {
+        return categoryApi.list(responseSpecUtil.OK_RESPONSE_SPEC).jsonPath()
+    }
+
+    JsonPath getOK(String categoryId) {
+        return categoryApi.get(categoryId, responseSpecUtil.OK_RESPONSE_SPEC).jsonPath()
+    }
+
+    String createOK(String body) {
+        JsonPath json = categoryApi.create(body, responseSpecUtil.OK_RESPONSE_SPEC).jsonPath()
+        return json.getString('id')
+    }
+
+    JsonPath deleteOK(String categoryId) {
+        return categoryApi.delete(categoryId, responseSpecUtil.OK_RESPONSE_SPEC).jsonPath()
+    }
+}
