--- src/integration-test/groovy/org/pih/warehouse/page/ChooseLocationPage.groovy
@@ -1,19 +1,0 @@
-package org.pih.warehouse.page
-
-import geb.Page
-
-class ChooseLocationPage extends Page {
-    static url = "dashboard/chooseLocation"
-
-    static at = {
-        title == 'Choose Location'
-    }
-
-    static content = {
-        locationButton(to: DashboardPage) { locationName -> $('a', 0, text: locationName) }
-    }
-
-    void chooseLocation(String locationName) {
-        locationButton(locationName).click()
-    }
-}
