+++ src/main/groovy/org/pih/warehouse/monitoring/SentryServletContainerInitializer.groovy
@@ -28,24 +28,24 @@ final class SentryServletContainerInitializer implements ServletContainerInitial
 
     @Override
     void onStartup(Set<Class<?>> c, ServletContext ctx) throws ServletException {
-        log.info 'Initializing sentry-servlet'
-        Properties properties = new Properties()
-        try {
-            InputStream is = this.class.classLoader.getResourceAsStream('git.properties')
-            properties.load(is)
-        } catch (Exception e) {
-            log.warn 'Unable to load git.properties file', e
-        }
-        Sentry.init {
-            it.with {
-                addInAppInclude('org.pih.warehouse')
-                attachServerName = true
-                enableAutoSessionTracking = true
-                // without this setting, SENTRY_DSN env. var is ignored
-                enableExternalConfiguration = true
-                release = properties.getProperty('git.commit.id', 'unknown')
-                sendDefaultPii = false
-            }
-        }
+//        log.info 'Initializing sentry-servlet'
+//        Properties properties = new Properties()
+//        try {
+//            InputStream is = this.class.classLoader.getResourceAsStream('git.properties')
+//            properties.load(is)
+//        } catch (Exception e) {
+//            log.warn 'Unable to load git.properties file', e
+//        }
+//        Sentry.init {
+//            it.with {
+//                addInAppInclude('org.pih.warehouse')
+//                attachServerName = true
+//                enableAutoSessionTracking = true
+//                // without this setting, SENTRY_DSN env. var is ignored
+//                enableExternalConfiguration = true
+//                release = properties.getProperty('git.commit.id', 'unknown')
+//                sendDefaultPii = false
+//            }
+//        }
     }
 }
